created a new ProductManagement page to handle all product-related endpoints for admin.
