+++ client/src/App.jsx
@@ -4,14 +4,16 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./components/Home.jsx";
 import About from "./components/About.jsx";
 import Checkout from "./components/CheckoutCart.jsx";
-import Navbar from "./components/Navbar.jsx";
-import UserAccount from "./components/UserAccount.jsx";
+import Navbar from "./components/navbar.jsx";
+import UserAccount from "./components/userAccount.jsx";
 import AdminAccount from "./components/AdminAccount.jsx";
 import UserHome from "./components/UserHome.jsx";
 import UserOrderHistory from "./components/UserOrderHistory.jsx";
 import Wishlist from "./components/Wishlist.jsx";
 import AllOrders from "./components/AllOrders.jsx";
 import AllUsers from "./components/AllUsers.jsx";
+import ProductManagement from "./components/ProductManagement.jsx";
+
 
 function App() {
   return (
@@ -30,7 +32,8 @@ function App() {
           <Route path="/orderhistory" element={<UserOrderHistory />} />
           <Route path="/wishlist" element={<Wishlist />} />
           <Route path="/allorders" element={<AllOrders />} />
-          <Route path="allusers" element={<AllUsers />} />
+          <Route path="/allusers" element={<AllUsers />} />
+          <Route path="/product-management" element={<ProductManagement />} />
         </Routes>
       </div>
     </BrowserRouter>
